+++ packages/eas-cli/src/fetch.ts
@@ -1,8 +1,15 @@
-import { Agent } from 'https';
-import createHttpsProxyAgent from 'https-proxy-agent';
-import fetch, { RequestInfo, RequestInit, Response } from 'node-fetch';
+import { env } from 'node:process';
+import {
+  ProxyAgent,
+  type RequestInfo,
+  type RequestInit,
+  type Response,
+  fetch,
+  getGlobalDispatcher,
+  setGlobalDispatcher,
+} from 'undici';
 
-export * from 'node-fetch';
+export { Agent, Headers, type RequestInfo, type RequestInit, Response } from 'undici';
 
 export class RequestError extends Error {
   constructor(
@@ -13,23 +20,21 @@ export class RequestError extends Error {
   }
 }
 
-function createHttpsAgent(): Agent | null {
-  const httpsProxyUrl = process.env.https_proxy;
-  if (!httpsProxyUrl) {
-    return null;
-  }
-  return createHttpsProxyAgent(httpsProxyUrl);
-}
-
-export const httpsProxyAgent: Agent | null = createHttpsAgent();
-
 export default async function (url: RequestInfo, init?: RequestInit): Promise<Response> {
-  const response = await fetch(url, {
-    ...init,
-    ...(httpsProxyAgent ? { agent: httpsProxyAgent } : {}),
-  });
+  installProxyAgent();
+
+  const response = await fetch(url, init);
   if (response.status >= 400) {
     throw new RequestError(`Request failed: ${response.status} (${response.statusText})`, response);
   }
+
   return response;
+}
+
+function installProxyAgent(): void {
+  const httpsProxyUrl = env.https_proxy;
+
+  if (httpsProxyUrl && !(getGlobalDispatcher() instanceof ProxyAgent)) {
+    setGlobalDispatcher(new ProxyAgent(httpsProxyUrl));
+  }
 }